pyproject.toml
[project]
name = "inacex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "multiprocess>=0.70.17",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "streamlit>=1.41.1",
    "uuid>=1.30",
    "uvicorn>=0.34.0",
]
